% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell_plot.R
\name{cell.plot}
\alias{cell.plot}
\title{Cell Plot}
\usage{
cell.plot(x, cells, lab.col = NULL, cell.col = c("blue", "white", "red"),
  inf.shading = 30/cell.lwd, space = 0.1, x.mar = c(0.2, 0.1),
  y.mar = c(0.08, 0), x.bound = NULL, lab.cex = 1, xdes.cex = 1,
  xlab.cex = 1, xlab.ticks = 5, xlab.yoffset = 0.08, sym = FALSE,
  cell.lwd = 2, cell.outer = 2, cell.sort = T, cell.limit = 50,
  cell.bounds = NULL, xlab = "", key = T, key.lab = "Color Key",
  key.n = 11, spacers = NULL, bar.scale = 1, ...)
}
\arguments{
\item{x}{Strictly positive numeric vector.}

\item{cells}{List of vectors. Must be the same length as x.}

\item{lab.col}{Vector of color names. Must be the same length as x. Defaults
to black.}

\item{cell.col}{Character vector of length 3, specifying colors for the low end, zero, and high
end of cell values. Defaults to c("blue","white","red").}

\item{inf.shading}{Numeric. Density of shading lines for infinite values. Defaults to 30/cell.lwd.}

\item{space}{Scaling factor for spacing between bars.}

\item{x.mar}{Numeric vector defining the inside margin to the left and right
of the main plot. Defaults to c(0.2,0.1).}

\item{y.mar}{Numeric vector defining the inside margin to the top and bottom
of the main plot. Defaults to c(0.08,0).}

\item{x.bound}{Numeric, must be positive. Specifies upper bound of x-axis scale. If NULL (the default),
this value is chosen automatically.}

\item{lab.cex}{Scaling factor for label text size.}

\item{xdes.cex}{Scaling factor for key legend text size.}

\item{xlab.cex}{Scaling factor for x-label text size.}

\item{xlab.ticks}{Number of ticks for the x-axis.}

\item{sym}{Logical, if \code{TRUE} visualize cell values on a symmetrical scale.}

\item{cell.lwd}{Size of the border of individual cells. Defaults to 1.}

\item{cell.outer}{Size of outer border around bars. Defaults to 2.}

\item{cell.sort}{Should cell values be sorted? Defaults to TRUE.}

\item{cell.limit}{Number of cells above which separators between cells are
omitted. Defaults to 50.}

\item{cell.bounds}{Numeric vector of length 2, specifying lower and upper
bound for cell value visualization. If set to NULL (the default), these
values are chosen automatically.}

\item{xlab}{Label for x-axis.}

\item{key.lab}{Label for color key.}

\item{key.n}{Number of legend boxes for the color key. Minimum is 3 (5 when
infinite values are present).}

\item{spacers}{Numeric vector. Inserts empty space after the specified
positions to visually group bars.}

\item{bar.scale}{Numeric. Optional multiplier for individual bar height. If the
resulting plot would exceed the plotting area, this value is scaled down to fit.}

\item{\dots}{Arguments passed through to the title() function. E.g.
main = "Plot title".}
}
\description{
Plots a horizontal barchart of strictly positive values in x. For each
entry, a vector of additional values needs to be provided in a list. The
additional values are plotted as cells subdividing the length of the
corresponding bar. A typical usage scenario is plotting the enrichment of
Gene Ontology terms, with individual cells reflecting the differential
regulation of the constituent genes.
}
\details{
Details
}
\examples{
\dontrun{
## Generate random positive vector and name it
x = sort( runif(16, min = 1, max = 3), decreasing = T )
names(x) = paste("GO Term",1:16)

## Label colors
xcolor = c(rep("darkslategrey",4), rep("chartreuse4",4), rep("coral4",8))

## Generate list with random vectors, one for each entry in x
cells = list()
xc = round( runif(16, min=21, max=100) )
for (i in 1:length(xc)) { cells = c(cells, list(runif(xc[i],-5,5))) }
cells[[9]][1:2] = Inf
cells[[9]][3] = -Inf

## Plot with spacers
cell.plot(x, cells, xcolor, spacers = c(4,8), xlab.ticks = 5,
  main="Cell Plot Demo", xlab="log(enrichment)", cell.limit = 80)
}
}
\author{
Robert Sehlke
}

