% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cellplot-package}
\alias{cellplot-package}
\title{Package cellplot}
\description{
Plot methods for presentation of cellular data as bar charts.
}
\details{
\code{\link{cell.plot}} - A function to plot segmented bars.\cr
\code{\link{golub.deg}} - A data set with differential gene expression
and gene ontology analysis derived from \link[multtest]{golub} data.\cr
}
\author{
Robert Sehlke\cr
Sven E. Templer
}

