% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cellplot}
\alias{cellplot}
\alias{cellplot-package}
\title{cell.plot

Plots a horizontal barchart of strictly positive values in x. For each entry, a vector of additional values needs to be provided in a list.
The additional values are plotted as cells subdividing the length of the corresponding bar.
A typical usage scenario is plotting the enrichment of Gene Ontology terms, with individual cells reflecting the differential regulation of the constituent genes.

\tabular{ll}{
Package: \tab cellplot\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-04-21\cr
License: \tab GPL-2\cr
}}
\description{
cell.plot

Plots a horizontal barchart of strictly positive values in x. For each entry, a vector of additional values needs to be provided in a list.
The additional values are plotted as cells subdividing the length of the corresponding bar.
A typical usage scenario is plotting the enrichment of Gene Ontology terms, with individual cells reflecting the differential regulation of the constituent genes.

\tabular{ll}{
Package: \tab cellplot\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-04-21\cr
License: \tab GPL-2\cr
}
}
\author{
RS
}

