% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/go.enrich.r
\name{go.enrich}
\alias{go.enrich}
\title{GO enrichment test and conversion of topGO data to list for cellplot}
\usage{
go.enrich(go, deg, col.stats = c("log2fc", "p.adj"), col.genes = "gene",
  n.top = NULL, test = GOFisherTest, name = "Fisher test", p.max = 0.05,
  p.adj.method = "fdr", order.by = c("enrichment", "pvalue"))
}
\arguments{
\item{go}{A \code{topGO} object.}

\item{deg}{A data frame with the differential gene expression statistics
and gene identifiers.}

\item{col.genes}{A character vector with the column name containing the
gene identifiers.}

\item{n.top}{An integer value referring to the number of top hits to extract.}

\item{test}{A \code{topGO} test function, e.g. \link[topGO]{GOFisherTest}.
Used as \code{testStatistic} argument value in the \link[topGO]{classicCount-class}
creator function.}

\item{name}{A character string for the test.
Used as \code{name} argument value in the \link[topGO]{classicCount-class}
creator function.}

\item{p.max}{A numeric value for the significance threshold (p value)
of the GO enrichment test. Set to \code{1} to extract all results from
the test.}

\item{p.adj.method}{A character string naming the p value adjustment
method. Use \code{"none"} to pass through. See \link{p.adjust}.}

\item{order.by}{A character vector, either \code{"e"} or \code{"p"} to sort
the output by enrichment or p values of the GO enrichment analysis.}

\item{col.stat}{A character vector with the column name(s) from \code{deg}
that contain statistics, e.g. log fold change or (adjusted) p values.}
}
\value{
The result is a \code{data.frame}. It contains the output of \code{GenTable}
from the \code{topGO} package merged with gene names and expression statistics
from a differential gene expression data frame.\cr
Columns are described at the \code{\link{golubstat}} data help page. Find an
example workflow in the examples section, too.\cr
The output is suitable for the \link{cell.plot} or \link{sym.plot} functions.\cr
}
\description{
The function performs the GO enrichment tests and merges
the gene expression statistics into the output. A \code{topGO} object
needs to be provided as input to run the enrichment test on, as well
as a corresponding data.frame containing the gene expression tests.
}
\author{
Sven E. Templer [aut]
}

