% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_histogram.R
\name{go.histogram}
\alias{go.histogram}
\title{GO Histogram}
\usage{
go.histogram(framelist, go.alpha = 0.05, gene.alpha = 0.05,
  go.alpha.term = "Elim", gene.alpha.term = "pvalue",
  logfc.term = "log2FoldChange", min.sig = 1, min.genes = 10,
  max.genes = 100, bar.scale = NULL, reorder = T,
  main = "GO enrichment", show.go.id = FALSE, prefix = "",
  show.ttest = F, lab.cex = 1, axis.cex = 1, group.cex = NULL,
  go.selection = NULL, term.selection = NULL, main.cex = 1)
}
\arguments{
\item{framelist}{List of TopGO data frames. The frames have to include two
additional columns: 'Upregulated' and 'Downregulated', with the numbers of
up- and down-regulated genes, respectively.}

\item{go.alpha}{Significance threshold for enrichment. Terms that are not
significantly enriched are still displayed using grey bars. Defaults to 0.05.}

\item{gene.alpha}{Significance threshold for the selection of genes. Defaults to 0.05.}

\item{go.alpha.term}{Identifies which column contains the p-values of the enrichment
analysis. Defaults to "Elim".}

\item{gene.alpha.term}{Identifies which column contains the p-values of the gene
differential expression analysis. Defaults to "pvalue".}

\item{min.sig}{Minimum times a term has to be detected as significantly
enriched across all groups to be displayed in the histogram. Defaults to 1.}

\item{min.genes}{Minimum number of genes for a significantly enriched term
to be included. Additional to the min.sig filter. Defaults to 10.}

\item{max.genes}{Maximum number of genes for a significantly enriched term
to be included. Defaults to 100.}

\item{bar.scale}{If not NULL (the default), set bar height to a fixed value relative to the
numeric factor provided. Use to ensure consistency across plots with varying numbers of elements.
Returns a warning if the plotting area is too small and uses the relative scale instead.}

\item{reorder}{If TRUE, terms significantly enriched in the same groups appear together.
Defaults to TRUE.}

\item{main}{Plot title.}

\item{show.go.id}{If TRUE, shows the GO ID number along with the term.}

\item{show.ttest}{If TRUE, displays p-values from one-sided t-tests beside the bars, coded as
asterisks. The p-values need to be included in the input data frames as columns "p.up.adj" 
and "p.down.adj" (NOTE: make more generic!). Defaults to FALSE.}

\item{lab.cex}{Scalar, Character expansion factor for labels.}

\item{axis.cex}{Scalar, Character expansion factor for axis tick labels.}

\item{group.cex}{Scalar, Character expansion factor for group labels.}

\item{go.selection}{(Optional) Character vector of GO IDs that should be displayed.}

\item{term.selection}{(Optional) Character vector of GO term identifiers that should be displayed.}

\item{main.cex}{Title character expansion.}
}
\description{
Plots a histogram, combining multiple topGO analyses into one figure. Bars
show the absolute numbers of up- and downregulated genes. Terms that were
detected as significantly enriched in the respective analysis are coloured,
grey bars indicate no enrichment. In addition, p-values from a one-sided
t-test on the significant log fold changes may be included, and are visualized
by asterisks at the end of the bars.
}
\examples{
\dontrun{

}

}
\author{
Robert Sehlke [aut]\cr
Sven E. Templer [ctb]
}

