% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/golub.degtest.r
\name{golub.degtest}
\alias{golub.degtest}
\title{Perform DEG and GO enrichment analysis (of \code{golub} data)}
\usage{
golub.degtest(e, i, a.frame = NULL)
}
\arguments{
\item{e}{Matrix with expression values as in \code{golub} data.
Genes per rows, samples per column.}

\item{i}{Logical vector as column/sample index for \code{e}, for
example the leukemia classes from \code{golub} data.}

\item{a.frame}{A \code{data.frame} which can contains annotation information
like gene names and which is included into the output.}
}
\value{
Returns a \code{data.frame}. See \link{golub.deg} for more
information.
}
\description{
Method used to analyse differential gene expression of the
\code{golub} data to create the \code{golub.deg} data.
}
\author{
Sven E. Templer [aut]
}

