% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeGOdeg}
\alias{mergeGOdeg}
\title{Merge topGO results with gene expression tables}
\usage{
mergeGOdeg(go, deg, map, go.go = "GO.ID", deg.gene = "gene",
  deg.stats = c("pvalue", "log2FoldChange"), map.go = "GO",
  map.gene = "gene", as.intersect = FALSE)
}
\arguments{
\item{go}{The topGO \code{GenTable()} data.frame.}

\item{deg}{The gene expression data.frame. See details.}

\item{map}{The mapping table for GO term to gene name conversion. See details.}

\item{go.go}{The column name for the GO identifiers.}

\item{deg.gene}{The column name for the gene identifiers.}

\item{deg.stats}{The column name(s) for the gene expression statistics
(e.g. p-values or log2 fold changes).}

\item{map.go}{The column name for the GO identifiers.}

\item{map.gene}{The column name for the gene identifiers.}
}
\description{
Merge a topGO result from the \code{GenTable} output with a gene expression table
to have a list of gene names and statistics for each GO term.
}
\details{
tba
}
\author{
Sven E. Templer [aut]
}

