% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topgo2cellplot.r
\name{topgo2cellplot}
\alias{topgo2cellplot}
\title{GO enrichment test and conversion of topGO data to list for cellplot}
\usage{
topgo2cellplot(go, deg.logfc = NULL, deg.genes = NULL, n = 30,
  ids = FALSE, replace.infinite = FALSE, p.adj.method = "fdr",
  topgo.test = GOFisherTest, topgo.test.name = "Fisher test")
}
\arguments{
\item{go}{A \code{topGO} object.}

\item{deg.logfc}{A numeric vector with the log fold changes from the
differential gene expression analysis.}

\item{deg.genes}{A character vector with the gene names corresponding
to \code{deg.logfc}.}

\item{n}{An integer value referring to the number of top hits to extract.}

\item{ids}{A logical value to attach GO IDs to the GO terms.}

\item{replace.infinite}{A logical value to replace infinite values
in \code{deg.logfc} with absolute maximums plus/minus one, depending on
the sign of the maximums.}

\item{p.adj.method}{A character string naming the p value adjustment
method. Use \code{"none"} to pass through. See \link{p.adjust}.}

\item{topgo.test}{A \code{topGO} test function, e.g. \link{GOFisherTest}.
Used as \code{testStatistic} argument value in the \link{classicCount-class}
creator function.}

\item{topgo.test.name}{A character string for the test.
Used as \code{name} argument value in the \link{classicCount-class}
creator function.}
}
\description{
The function performs the GO enrichment tests (a Fisher test).
It extracts data from \code{topGO} objects and combines it with
differential gene expression fold changes to create a data list
suitable for \code{cellplot} input.
}
\author{
Sven E. Templer [aut]
}

