% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arc_plot.R
\name{arc.plot}
\alias{arc.plot}
\title{Arc Plot}
\usage{
arc.plot(x, up.list, down.list, y.mar = c(0, 0), x.mar = c(0.5, 0),
  x.bound = NULL, x.ticks = 3, x.scale = 1, space = 0.1,
  fixed.scale = NULL, t = 0.2, tc = "black", main = "GO Term Analysis",
  barcol = c("deepskyblue2", "coral"))
}
\arguments{
\item{x}{Strictly positive numeric vector.}

\item{up.list}{List with named vectors, according to x, but only with positive
(i.e. fold changes of gene expression) values.}

\item{down.list}{As up.list, but with negatives.}

\item{y.mar}{Plot area horizontal margins.}

\item{x.mar}{Plot area vertical margins.}

\item{x.bound}{tba}

\item{x.ticks}{tba}

\item{x.scale}{tba}

\item{space}{tba}

\item{fixed.scale}{tba}

\item{t}{Threshold for Jaccard distance to display.}

\item{tc}{Color of fold change bars.}

\item{main}{Plot title.}

\item{barcol}{Color of ratio bars.}

\item{...}{tba}
}
\description{
Plots a combined barchart for a differential expression experiment GO analysis.
Shows enrichment score (or similar), proportions of up- and downregulated set members, as well as
set overlap. Bar heights correspond to the number of genes of interest in the term. Overlap between
terms is visualized by an arc connecting them. The line width of the arc corresponds to the number
of overlapping genes, on the same scale as bar height. Arc colour scale maps to the Jaccard Index
( |Overlap| / |Union| ), bounded between 0 and 1, where black represents maximum overlap, i.e. Jaccard
Index of close or equal to 1. A threshold on the Jaccard index may be set to suppress arcs below the
specified value. Vertical ordering of terms is the result of hierarchical clustering, using hclust on
the pairwise Jaccard Distance matrix (1 - Jaccard Index).
}
\examples{
\dontrun{

### golub.deg data example
x = subset(golubstat, p<=.05 & significant>4 & !duplicated(genes))
x = head(x, 10)

lfcvec = x$deg.log2fc
genevec = x$genes

uplist = list()
downlist = list()
for (i in 1:length(lfcvec)) {
  names(lfcvec[[i]]) = genevec[[i]]
  uplist[[i]] = lfcvec[[i]][lfcvec[[i]] > 0]
  downlist[[i]] = lfcvec[[i]][lfcvec[[i]] < 0]
}

arc.plot( x = setNames(x$loge, x$term), up.list = uplist, down.list = downlist)
}
}
\author{
Robert Sehlke [aut]\cr
Sven E. Templer [ctb]
}

