\name{cellplot-package}
\alias{cellplot-package}
\alias{cellplot}
\docType{package}
\title{
Cell Plot, segmented bar charts.
}
\description{
Plots horizontal bars that are subdivided into cells, allowing the mapping of additional information. A typical usage scenario is plotting the enrichment of GO terms, with cell colors encoding the logfold-changes of individual annotated genes.
}
\details{
\tabular{ll}{
Package: \tab cellplot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-21\cr
License: \tab GPL-2\cr
}
}
\author{
Author and maintainer: Robert Sehlke <rsehlke@age.mpg.de>
}
\references{
-
}
\keyword{ package }
\seealso{
\code{\link[cellplot:cell.plot]{cell.plot()}}
}
\examples{
See cell.plot() manual page.
}
