% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CellPlot-package}
\alias{CellPlot}
\alias{CellPlot-package}
\title{Package CellPlot}
\description{
Plot methods for presentation of cellular data as bar charts.
}
\details{
Plot functions:\cr
\code{\link{cell.plot}} - A function to plot segmented bars.\cr
\code{\link{sym.plot}} - A function to plot split bars.\cr
\code{\link{arc.plot}} - A function to plot category enrichments
showing proportions of member directionalities and .\cr\cr
Analysis functions:\cr
\code{\link{go.enrich}} - A function to merge \code{topGO} objects
and differential gene expression tables and perform GO enrichment
analysis returning a \code{data.frame} suitable for the
\code{cell.plot} and \code{sym.plot} functions.\cr\cr
Data objects:\cr
\code{\link{golubstat}} - A data set with differential gene expression
and gene ontology analysis derived from \link[multtest]{golub} data.\cr
}
\author{
Robert Sehlke\cr
Sven E. Templer
}

